requirements.txt
streamlit
python-chess
firebase-admin